# server: make CI faster