--- a/server/server_test.go
+++ b/server/server_test.go
@@ -40,6 +40,7 @@ import (
 	"github.com/pingcap/tidb/config"
 	"github.com/pingcap/tidb/errno"
 	"github.com/pingcap/tidb/kv"
+	"github.com/pingcap/tidb/session"
 	"github.com/pingcap/tidb/util/logutil"
 	"github.com/pingcap/tidb/util/versioninfo"
 	"github.com/tikv/client-go/v2/tikv"
@@ -59,12 +60,8 @@ func TestT(t *testing.T) {
 	}
 
 	// AsyncCommit will make DDL wait 2.5s before changing to the next state.
-	// In the testing environment, there is no correctness issue. Waiting is unnecessary, it just makes CI slow.
-	// So set the config to avoid that.
-	config.UpdateGlobal(func(conf *config.Config) {
-		conf.TiKVClient.AsyncCommit.SafeWindow = 0
-		conf.TiKVClient.AsyncCommit.AllowedClockDrift = 0
-	})
+	// Set schema lease to avoid it from making CI slow.
+	session.SetSchemaLease(0)
 	CustomVerboseFlag = true
 	logLevel := os.Getenv("log_level")
 	err := logutil.InitLogger(logutil.NewLogConfig(logLevel, logutil.DefaultLogFormat, "", logutil.EmptyFileLogConfig, false))