@tisonkun 2021-07-26
> #26289 is a better solution, under that limitation, if a test case can not finish within 3s, the author should optimize it, otherwise move it out of the unit test. Maybe a integration test is better place for such case.
> 
> @zhouqiang-cl @tisonkun
> 
> Even better if we can monitor the CI time cost and have a dashboard for it, so we can see how it changes.

Yep. I post a script that we might build a fixture for timeout under #26289. For report, I'd suggest add a ci task to generate from `go test -json`.

[JUnit](https://javabydeveloper.com/junit-5-timeouts-fail-if-test-not-executed-within-time/) provides a nice example which Golang ecosystem lack of.